requirements.txt
numpy
pandas
scikit-learn
pyyaml
joblib
streamlit
lightgbm
matplotlib
seaborn
